refactor: 공통 컴포넌트 및 함수는 shared 폴더에 넣기
