+++ features/application_detail/components/AddSchedule.tsx
@@ -1,9 +1,9 @@
-import { Card, CardContent, CardFooter } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
+import { Card, CardContent, CardFooter } from "@/shared/components/ui/card";
+import { Input } from "@/shared/components/ui/input";
+import { Label } from "@/shared/components/ui/label";
 import { DatePickerInput } from "./DatePickerInput";
 import StatusButtonGroup from "./StatusButtonGroup";
-import { Button } from "@/components/ui/button";
+import { Button } from "@/shared/components/ui/button";
 import { RotateCcw, Save } from "lucide-react";
 
 export default function AddSchedule({setOpen}: {setOpen: React.Dispatch<React.SetStateAction<boolean>>}) {
