+++ features/application_detail/components/DatePickerInput.tsx
@@ -1,5 +1,5 @@
 import { DateTimePicker } from "./DateTimePicker"
-import { Label } from "@/components/ui/label"
+import { Label } from "@/shared/components/ui/label"
 
 interface DatePickerInputProps {
   value?: string
