+++ features/application_detail/components/DocumentStage/ViewCard.tsx
@@ -1,4 +1,4 @@
-import { Label } from "@/components/ui/label";
+import { Label } from "@/shared/components/ui/label";
 import { Clock } from "lucide-react";
 import { StatusButton } from "../StatusButton";
