+++ features/application_detail/components/StageDetailNoteCard/EditCard.tsx
@@ -1,7 +1,7 @@
-import { Button } from "@/components/ui/button";
-import { CardContent } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
+import { Button } from "@/shared/components/ui/button";
+import { CardContent } from "@/shared/components/ui/card";
+import { Input } from "@/shared/components/ui/input";
+import { Textarea } from "@/shared/components/ui/textarea";
 import { Globe, MapPin, Paperclip, Upload, X } from "lucide-react";
 import FormLabel from "../common/FormLabel";
