+++ features/application_detail/components/common/StageEditButton.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@/components/ui/button";
+import { Button } from "@/shared/components/ui/button";
 import { Pen } from "lucide-react";
 
 interface StageEditButtonProps {
