+++ shared/components/application/ApplicationCard.tsx
@@ -1,8 +1,8 @@
 // JobApplicationCard.tsx (React Component)
-import { JobApplication } from '@/types';
+import { JobApplication } from '@/shared/types';
 import React from 'react';
-import { Card, CardHeader } from '@/components/ui/card';
-import { Badge } from '@/components/ui/badge';
+import { Card, CardHeader } from '@/shared/components/ui/card';
+import { Badge } from '@/shared/components/ui/badge';
 
 interface JobApplicationCardProps {
   children?: React.ReactNode;
